Refactor: 채팅 로직 모듈화
